variscite-packagegroup: add grep
grep is required for internal testing.
The "full" version is needed as we need Perl Regular Expressions which
busybox cannot provide.

Previously grep was provided as a dependency from pm-utils, but with
newer releases of busybox, the hard requirement for GNU grep was dropped
from pm-utils which meant that it was no longer automatically installed.

See here for details:
https://git.yoctoproject.org/poky/commit/meta/recipes-bsp/pm-utils/pm-utils_1.4.1.bb?id=795420850bd0bd2b0fd7d7c011e4ce346b448604

Signed-off-by: Matthias Proske <matthias.p@variscite.com>